Revamp HTTP/1 encoder: - Combine the 2 state machines into 1. - Encoder no longer deals with the lock when getting new chunks. The stream will be responsible for getting new chunks into the thread.

+++ include/aws/http/request_response.h
@@ -319,7 +319,7 @@ struct aws_http1_chunk_options {
     /*
      * Size of the chunk_data input stream in bytes.
      */
-    size_t chunk_data_size;
+    uint64_t chunk_data_size;
 
     /**
      * A pointer to an array of chunked extensions.
